Cmake: Refactor dependency fetching
- Build maiatrac3plus on targets without ffmpeg
