+++ cli/CMakeLists.txt
@@ -66,7 +66,7 @@ install(TARGETS vgmstream_cli
 
 # TODO: Make it so vgmstream123 can build with Windows (this probably needs a libao.dll included with vgmstream, though)
 
-if(NOT WIN32 AND USE_V123)
+if(NOT WIN32 AND BUILD_V123)
 	# vgmstream123
 
 	add_executable(vgmstream123
@@ -75,7 +75,7 @@ if(NOT WIN32 AND USE_V123)
 	# Link to the vgmstream library as well as libao
 	target_link_libraries(vgmstream123
 		libvgmstream
-		ao dl)
+		ao pthread dl)
 
 	setup_target(vgmstream123 TRUE)
