+++ src/coding/at3plus_decoder.h
@@ -1,0 +1,11 @@
+#ifndef _AT3PLUS_DECODER_H
+#define _AT3PLUS_DECODER_H
+
+struct maiatrac3plus_codec_data {
+    sample_t* buffer;
+    int channels;
+    int samples_discard;
+    void* handle;
+};
+
+#endif
